third_party/blink/web_tests/storage/indexeddb/createObjectStore-name-argument-required-expected.txt

Test IndexedDB createObjectStore name argument is required

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

dbname = "createObjectStore-name-argument-required.html"
indexedDB.deleteDatabase(dbname)
indexedDB.open(dbname)
db = event.target.result
PASS db.createObjectStore(); threw exception TypeError: Failed to execute 'createObjectStore' on 'IDBDatabase': 1 argument required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE
